requirements.txt
langchain==0.0.235
python-dotenv==1.0.0
streamlit==1.24.0
openai==0.27.8
